requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
drf-spectacular[sidecar]>=0.27,<0.28
djangorestframework-csv>=3.0,<3.1